refactor: move getCSSVarValue to shared utils
getCSSVarValue is now exported and able to be used as util in other apps.
