+++ packages/appChrome/components/SidebarSubMenuItem.tsx
@@ -11,7 +11,7 @@ import { textSize, flex, tintContent } from "../../shared/styles/styleUtils";
 import Clickable from "../../clickable/components/clickable";
 import ResetLink from "../../link/components/ResetLink";
 import { useTheme } from "@emotion/react";
-import getCSSVarValue from "../../utilities/getCSSVarValue";
+
 import {
   themeTextColorPrimary,
   themeTextColorPrimaryInverted,
@@ -22,6 +22,7 @@ import {
 import { getTextColor } from "../../shared/styles/color";
 import { AppChromeTheme, SidebarNavItemProps } from "../types";
 import { IconSize } from "../../shared/types/iconSize";
+import { getCSSVarValue } from "../../shared/styles/styleUtils/typography/color";
 
 export interface SidebarSubMenuItemTextProps {
   menuHasIcon: boolean;
