+++ packages/donutChart/components/DonutChart.tsx
@@ -3,9 +3,11 @@ import {
   themeBorder,
   themeBrandPrimary
 } from "../../design-tokens/build/js/designTokens";
-import { tintContentSecondary } from "../../shared/styles/styleUtils/typography/color";
+import {
+  getCSSVarValue,
+  tintContentSecondary
+} from "../../shared/styles/styleUtils/typography/color";
 import { hexToRgbA } from "../../shared/styles/color";
-import getCSSVarValue from "../../utilities/getCSSVarValue";
 
 interface DonutChartDatum {
   percentage: number;
