+++ packages/shared/styles/styleUtils/typography/color.ts
@@ -45,3 +45,34 @@ export const tintContentSecondary = css`
     ${tintContent(themeTextColorSecondaryInverted)};
   }
 `;
+
+/**
+ * This function retrieves the value of a CSS variable
+ * from the document root or a provided fallback value.
+ *
+ * @param {string} cssVar - The CSS variable in the format of "var(--variable-name, fallback-value)"
+ *
+ * @returns {string} - Returns the computed value of the CSS variable if it's defined in the document root.
+ *
+ * If the variable is not defined or the function is called outside of a browser context, it returns the fallback value provided.
+ */
+export const getCSSVarValue = cssVar => {
+  // Get the content between the parentheses in "var()"
+  // and split the args from a string to an array
+  const cssVarArgsString = cssVar.match(/\(([^)]+)\)/)[1];
+  const cssVarArgs = cssVarArgsString.replace(/\s/g, "").split(",");
+
+  if (typeof window === "undefined") {
+    return cssVarArgs[1];
+  }
+
+  if (document?.documentElement !== null) {
+    // If there's a custom property defined on :root (<html>), return that.
+    // Otherwise, return the fallback value.
+    return (
+      getComputedStyle(document.documentElement)
+        .getPropertyValue(cssVarArgs[0])
+        .trim() || cssVarArgs[1]
+    );
+  }
+};
