+++ packages/utilities/index.ts
@@ -1,3 +1,2 @@
 export { default as findNestedPropertyInObject } from "./findNestedPropertyInObject";
-export { default as getCSSVarValue } from "./getCSSVarValue";
 export { default as resizeEventManager } from "./resizeEventManager";
